+++ activerecord/test/cases/adapters/postgresql/active_schema_test.rb
@@ -82,6 +82,16 @@ class PostgresqlActiveSchemaTest < ActiveRecord::PostgreSQLTestCase
       assert_equal expected, add_index(:people, :last_name, nulls_not_distinct: true)
     end
 
+    expected = %(CREATE INDEX "index_people_on_last_name" ON "people" ("last_name") WITH (fillfactor = 70))
+    assert_equal expected, add_index(:people, :last_name, with: { fillfactor: 70 })
+
+    expected = %(CREATE INDEX "index_people_on_last_name" ON "people" ("last_name") WITH (fillfactor = 70, deduplicate_items = on))
+    assert_equal expected, add_index(:people, :last_name, with: { fillfactor: 70, deduplicate_items: :on })
+
+    assert_raise ArgumentError do
+      add_index(:people, :last_name, with: "fillfactor = 70")
+    end
+
     assert_raise ArgumentError do
       add_index(:people, :last_name, algorithm: :copy)
     end